[AUD-1331] Move Explore page state to common pages part of the store
