+++ src/pages/explore-page/ExploreCollectionsPage.tsx
@@ -2,13 +2,13 @@ import React from 'react'
 
 import { connect } from 'react-redux'
 
+import { ExploreCollectionsVariant } from 'common/store/pages/explore/types'
 import { AppState } from 'store/types'
 import { isMobile } from 'utils/clientUtil'
 
 import ExploreCollectionsPageProvider from './ExploreCollectionsPageProvider'
 import DesktopCollectionsPage from './components/desktop/CollectionsPage'
 import MobileCollectionsPage from './components/mobile/CollectionsPage'
-import { ExploreCollectionsVariant } from './store/types'
 
 type OwnProps = {
   variant: ExploreCollectionsVariant
