+++ src/pages/explore-page/collections.ts
@@ -4,6 +4,7 @@ import { ReactComponent as IconCassette } from 'assets/img/iconCassette.svg'
 import { ReactComponent as IconExploreDJ } from 'assets/img/iconExploreDJ.svg'
 import { ReactComponent as IconExploreTopAlbums } from 'assets/img/iconExploreTopAlbums.svg'
 import { ReactComponent as IconExploreTopPlaylists } from 'assets/img/iconExploreTopPlaylists.svg'
+import { ExploreCollectionsVariant } from 'common/store/pages/explore/types'
 import {
   EXPLORE_LET_THEM_DJ_PAGE,
   EXPLORE_TOP_ALBUMS_PAGE,
@@ -11,8 +12,6 @@ import {
   TRENDING_PLAYLISTS_PAGE,
   TRENDING_UNDERGROUND_PAGE
 } from 'utils/route'
-
-import { ExploreCollectionsVariant } from './store/types'
 
 export type ExploreCollection = {
   variant: ExploreCollectionsVariant
