+++ src/store/reducers.js
@@ -19,8 +19,6 @@ import dashboard from 'pages/artist-dashboard-page/store/reducer'
 import collection from 'pages/collection-page/store/reducer'
 import deactivateAccount from 'pages/deactivate-account-page/store/slice'
 import deleted from 'pages/deleted-page/store/slice'
-import exploreCollections from 'pages/explore-page/store/collections/slice'
-import explore from 'pages/explore-page/store/reducer'
 import favorites from 'pages/favorites-page/store/reducers'
 import feed from 'pages/feed-page/store/reducer'
 import followers from 'pages/followers-page/store/reducers'
@@ -129,8 +127,6 @@ const createRootReducer = routeHistory =>
       }),
       pages: combineReducers({
         deleted,
-        explore,
-        exploreCollections,
         favorites,
         followers,
         following,
