+++ src/store/sagas.js
@@ -5,6 +5,8 @@ import coreCacheSagas from 'common/store/cache/sagas'
 import tracksSagas from 'common/store/cache/tracks/sagas'
 import usersSagas from 'common/store/cache/users/sagas'
 import errorSagas from 'common/store/errors/sagas'
+import exploreCollectionsPageSagas from 'common/store/pages/explore/exploreCollections/sagas'
+import explorePageSagas from 'common/store/pages/explore/sagas'
 import remoteConfigSagas from 'common/store/remote-config/sagas'
 import shareModalSagas from 'common/store/ui/share-modal/sagas'
 import addToPlaylistSagas from 'components/add-to-playlist/store/sagas'
@@ -22,8 +24,6 @@ import rewardsPageSagas from 'pages/audio-rewards-page/store/sagas'
 import collectionSagas from 'pages/collection-page/store/sagas'
 import deactivateAccountSagas from 'pages/deactivate-account-page/store/sagas'
 import deletedSagas from 'pages/deleted-page/store/sagas'
-import exploreCollectionsPageSagas from 'pages/explore-page/store/collections/sagas'
-import explorePageSagas from 'pages/explore-page/store/sagas'
 import favoritePageSagas from 'pages/favorites-page/store/sagas'
 import feedPageSagas from 'pages/feed-page/store/sagas'
 import followersPageSagas from 'pages/followers-page/store/sagas'
